rinha/requirements.txt
agate==1.6.3
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
Babel==2.10.1
botocore==1.29.8
certifi==2024.2.2
cffi==1.15.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
dbt-core==1.1.0
dbt-extractor==0.4.1
dbt-postgres==1.1.0
docutils==0.16
exceptiongroup==1.2.0
fastapi==0.109.2
future==1.0.0
h11==0.14.0
hologram==0.0.14
idna==3.6
importlib-metadata==7.0.1
isodate==0.6.1
Jinja2==2.11.3
jmespath==1.0.1
jsonschema==3.1.1
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.0.1
mashumaro==2.9
minimal-snowplow-tracker==0.0.2
msgpack==1.0.3
networkx==2.8
packaging==21.3
parsedatetime==2.4
psycopg==3.1.18
psycopg2-binary==2.9.3
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.36.3
text-unidecode==1.3
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.27.1
Werkzeug==2.1.2
youtube-dl==2021.12.17
zipp==3.17.0
